Branch AssetTransfer into Automatic & Manual (nested) subclasses
Try to ensure that 'AssetTransfer::sendFunds' has a consistent meaning
of triggering the sending of funds, either automatically via a wallet
bridge or by prompting the user to start a manual transfer. Accordingly,
do not call 'sendFunds' from 'MakerMultiSigProtocol::onFundsSent', but
instead call it earlier from 'onDepositTxConfirmed'.

To move the logic for triggering the manual user payment prompt into
'AssetTransfer.Manual', provide the subclass with a listener class to
attach a user prompt hook with, plus an 'onManualPaymentStarted' event
method. Also add the two new payment specific MultiSigProtocol states,
'START_MANUAL_PAYMENT' & 'MANUAL_PAYMENT_STARTED', so that (for now) the
user action can continue to be triggered and waited on via a pair of
protocol state changes.