Add ManyParty(Contract|Protocol) & rename CounterParty to Party
Provide 'ManyPartyContract' & 'ManyPartyProtocol' classes in addition to
the current 2 party classes, extending Contract & Protocol respectively.
This is to support 3-or-more party protocols such as a Lightning escrow
trade protocol, which requires a 3rd party escrow agent with whom HTLCs
(pending payments) are set up at the start of the trade (such that the
escrow times out and defaults to him if and only if there is a dispute
between the two traders).

Also rename 'CounterParty' to 'Party', as the new Contract & Protocol
subclasses hold a map from all roles to parties, including oneself, not
just a lone counterparty.