Add initial shared state framework & Lightning escrow protocol
Provide a set of custom type & method annotations,

  @State, @DependsOn, @Action, @Supplied, @Event, @Access,

for annotating an interface of trade state properties (getter methods),
declaring the interdependencies between them, when and for whom they are
available, plus the actions which are triggered by their availability.

Provide a class, 'SharedStateFactory', to process the annotations at
runtime and create an associated 'java.lang.reflect.Proxy' instance to
hold the corresponding trade state fields. The proxy will accept/produce
messages passed from/to the other contract parties, updating the
relevant fields and invoking the relevant actions. (The proxy
construction is not done yet - just the annotation processing for now.)

Also provide a preliminary Lightning escrow trade protocol to exercise
the framework, with associated interface 'LightningEscrow.SharedState'
using the new annotations. (The protocol is non-functional for now.)