Add initial SharedState proxy construction
Provide a 'factory.create()' method which uses 'java.lang.reflect.Proxy'
to instantiate an annotated shared state interface. The proxy holds a
map, of type 'Map<String, FieldVersions<Object>>', from property names
to all the different versions of each initialised property. As a
property can have multiple '@DependsOn' annotations and may be
additionally supplied externally via an incoming message or during
initialisation of the trade state, the framework will attempt to keep
track of the different versions and their sources to make sure that they
are all the same. For example, 'MultiSig.SharedState::unsignedDepositTx'
includes the '@DependsOn' alternatives:

  {'makerDepositTxInputs', 'takerDepositTxInputs', 'escrowAddress'},
  {'takerSignedDepositTx'},

so the unsigned deposit tx may be constructed either from the initial
inputs or by stripping the signature from the signed deposit tx received
from the taker. If these are not byte-for-byte the same, then something
has gone wrong and the framework will be able to detect that (TODO).

To the above end, add a 'FieldVersions' class to hold all the available
property versions and their sources. Also add an 'IndexInjector' utility
class to allow the framework to select the appropriate '@DependsOn'
alternative to use, via the 'oneOf(..)' selector called by the default
method of the corresponding property. The return value is then memoised.

Also add rudimentary logic to fire events and send/receive messages
from/to the proxy object. Use an eager strategy of messaging every field
to each counterparty which has access to it, as soon as it is learned
by the state owner (as long as the same field hasn't already been sent
to or received from the given counterparty).